refactor: 분리한 리듀서 적용 및 데이터 비동기 작업 에러 로직 추가
