+++ src/constants/message.js
@@ -1,4 +1,6 @@
 const MESSAGE = {
+  FAIL_FETCH_DATA: '데이터를 불러오는데 실패했습니다.',
+
   SUCCESS_ADD_ITEM_TO_CART: '장바구니에 추가되었습니다.',
   FAIL_ADD_ITEM_TO_CART: '장바구니에 추가하지 못했습니다.',
