+++ src/pages/Order/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { useHistory } from 'react-router';
 import { useDispatch, useSelector } from 'react-redux';
-import { addOrderDetail } from '../../store';
 import { deleteCartItems } from '../../store/cartReducer';
+import { addOrderDetail } from '../../store/orderListReducer';
 import { API } from '../../services'
 import { Button, HighlightText, Product } from '../../components/shared';
 import { COLOR, MESSAGE, PATH } from '../../constants';
